add Piece bounds constriction, change Grid to contain Uint32Array, create TetrisGrid

+++ tests/Grid.createRotated.test.ts
@@ -1,7 +1,7 @@
 import {Grid} from "../src/Grid";
 
 test("Grid get angle 0", () => {
-    const grid = new Grid(5, 4, new Uint8Array([
+    const grid = new Grid(5, 4, new Uint32Array([
         4, 0, 1, 0,
         3, 0, 1, 0,
         0, 0, 1, 0,
@@ -9,10 +9,8 @@ test("Grid get angle 0", () => {
         2, 0, 0, 1,
     ]));
 
-
-
     const rotated = grid.createRotated(0);
-    expect(rotated.grid).toEqual(new Uint8Array([
+    expect(rotated.grid).toEqual(new Uint32Array([
         4, 0, 1, 0,
         3, 0, 1, 0,
         0, 0, 1, 0,
@@ -22,7 +20,7 @@ test("Grid get angle 0", () => {
 });
 
 test("Grid get angle 1", () => {
-    const grid = new Grid(5, 4, new Uint8Array([
+    const grid = new Grid(5, 4, new Uint32Array([
         4, 0, 1, 0,
         3, 0, 1, 0,
         0, 0, 1, 0,
@@ -31,7 +29,7 @@ test("Grid get angle 1", () => {
     ]));
 
     const rotated = grid.createRotated(1);
-    expect(rotated.grid).toEqual(new Uint8Array([
+    expect(rotated.grid).toEqual(new Uint32Array([
         2, 0, 0, 3, 4,
         0, 0, 0, 0, 0,
         0, 1, 1, 1, 1,
@@ -40,7 +38,7 @@ test("Grid get angle 1", () => {
 });
 
 test("Grid get angle 2", () => {
-    const grid = new Grid(5, 4, new Uint8Array([
+    const grid = new Grid(5, 4, new Uint32Array([
         4, 0, 1, 0,
         3, 0, 1, 0,
         0, 0, 1, 0,
@@ -49,7 +47,7 @@ test("Grid get angle 2", () => {
     ]));
 
     const rotated = grid.createRotated(2);
-    expect(rotated.grid).toEqual(new Uint8Array([
+    expect(rotated.grid).toEqual(new Uint32Array([
         1, 0, 0, 2,
         0, 1, 0, 0,
         0, 1, 0, 0,
@@ -59,7 +57,7 @@ test("Grid get angle 2", () => {
 });
 
 test("Grid get angle 3", () => {
-    const grid = new Grid(5, 4, new Uint8Array([
+    const grid = new Grid(5, 4, new Uint32Array([
         4, 0, 1, 0,
         3, 0, 1, 0,
         0, 0, 1, 0,
@@ -68,7 +66,7 @@ test("Grid get angle 3", () => {
     ]));
 
     const rotated = grid.createRotated(3);
-    expect(rotated.grid).toEqual(new Uint8Array([
+    expect(rotated.grid).toEqual(new Uint32Array([
         0, 0, 0, 0, 1,
         1, 1, 1, 1, 0,
         0, 0, 0, 0, 0,
